+++ lib/flutter_file_info.dart
@@ -4,7 +4,8 @@ export 'src/file_info.dart';
 export 'src/models/icon_info.module.dart';
 export 'src/models/file_metadata.module.dart';
 
-export 'src/windows/enum/file_attribiutes.dart';
+export 'src/windows/enum/windows_file_attribiutes.dart';
+export 'src/android/enum/android_file_attribiutes.dart';
 
 export 'src/android/file_info_android.dart' show FileInfoAndroid;
 export 'src/windows/file_info_windows.dart' show FileInfoWindows;