+++ lib/src/models/file_metadata.module.dart
@@ -32,10 +32,19 @@ class FileMetadata extends Equatable {
   /// The Windows file attributes represented as a numerical bitmask.
   final int? dwFileAttributes;
 
-  /// The list of [FileAttributes] representing the file's attributes.
-  final List<FileAttributes>? attributes;
+  /// The list of Windows file attributes.
+  final List<WindowsFileAttributes>? winAttributes;
 
-  /// Creates a new instance of [FileMetadata] with the specified parameters.
+  /// The list of Android file attributes.
+  final List<AndroidFileAttributes>? androidAttributes;
+
+  /// Represents the metadata of a file.
+  ///
+  /// This class contains information about the attributes of a file.
+  /// The [winAttributes] property is a list of [WindowsFileAttributes] that represents
+  /// the Windows-specific attributes of the file.
+  /// The [androidAttributes] property is a list of [AndroidFileAttributes] that represents
+  /// the Android-specific attributes of the file.
   const FileMetadata({
     required this.filePath,
     this.fileName,
@@ -47,12 +56,13 @@ class FileMetadata extends Equatable {
     this.sizeBytes,
     this.fileSize,
     this.dwFileAttributes,
-    this.attributes,
+    this.winAttributes,
+    this.androidAttributes,
   });
 
   /// Returns a string representation of the [FileMetadata] object.
   @override
-  String toString() => 'FileInfo('
+  String toString() => 'FileMetadata('
       'filePath: $filePath, '
       'fileName: $fileName, '
       'fileExtension: $fileExtension, '
@@ -63,7 +73,8 @@ class FileMetadata extends Equatable {
       'sizeBytes: $sizeBytes, '
       'fileSize: $fileSize, '
       'dwFileAttributes: $dwFileAttributes, '
-      'attributes: $attributes)';
+      'winAttributes: $winAttributes, '
+      'androidAttributes: $androidAttributes)';
 
   /// Creates a new [FileMetadata] object with updated values based on the provided parameters.
   ///
@@ -79,7 +90,8 @@ class FileMetadata extends Equatable {
     int? sizeBytes,
     String? fileSize,
     int? dwFileAttributes,
-    List<FileAttributes>? attributes,
+    List<WindowsFileAttributes>? winAttributes,
+    List<AndroidFileAttributes>? androidAttributes,
   }) {
     return FileMetadata(
       filePath: filePath ?? this.filePath,
@@ -92,7 +104,8 @@ class FileMetadata extends Equatable {
       sizeBytes: sizeBytes ?? this.sizeBytes,
       fileSize: fileSize ?? this.fileSize,
       dwFileAttributes: dwFileAttributes ?? this.dwFileAttributes,
-      attributes: attributes ?? this.attributes,
+      winAttributes: winAttributes ?? this.winAttributes,
+      androidAttributes: androidAttributes ?? this.androidAttributes,
     );
   }
 
@@ -109,6 +122,7 @@ class FileMetadata extends Equatable {
         sizeBytes,
         fileSize,
         dwFileAttributes,
-        attributes,
+        winAttributes,
+        androidAttributes,
       ];
 }