+++ test/src/models/file_metadata.module_test.dart
@@ -16,7 +16,10 @@ void main() {
         sizeBytes: 1024,
         fileSize: '1 KB',
         dwFileAttributes: FILE_FLAGS_AND_ATTRIBUTES.FILE_ATTRIBUTE_READONLY,
-        attributes: const [FileAttributes.READ_ONLY, FileAttributes.HIDDEN],
+        winAttributes: const [
+          WindowsFileAttributes.READ_ONLY,
+          WindowsFileAttributes.HIDDEN
+        ],
       );
 
       expect(fileInfo.filePath, 'C:/test/file.txt');
@@ -30,8 +33,8 @@ void main() {
       expect(fileInfo.fileSize, '1 KB');
       expect(fileInfo.dwFileAttributes,
           FILE_FLAGS_AND_ATTRIBUTES.FILE_ATTRIBUTE_READONLY);
-      expect(fileInfo.attributes,
-          [FileAttributes.READ_ONLY, FileAttributes.HIDDEN]);
+      expect(fileInfo.winAttributes,
+          [WindowsFileAttributes.READ_ONLY, WindowsFileAttributes.HIDDEN]);
     });
 
     test('FileInfo copyWith creates a new instance with updated values', () {
@@ -46,14 +49,17 @@ void main() {
         sizeBytes: 1024,
         fileSize: '1 KB',
         dwFileAttributes: FILE_FLAGS_AND_ATTRIBUTES.FILE_ATTRIBUTE_READONLY,
-        attributes: const [FileAttributes.READ_ONLY],
+        winAttributes: const [WindowsFileAttributes.READ_ONLY],
       );
 
       final updatedFileInfo = fileInfo.copyWith(
         fileName: 'new_file.txt',
         sizeBytes: 2048,
         dwFileAttributes: FILE_FLAGS_AND_ATTRIBUTES.FILE_ATTRIBUTE_HIDDEN,
-        attributes: [FileAttributes.HIDDEN, FileAttributes.SYSTEM],
+        winAttributes: [
+          WindowsFileAttributes.HIDDEN,
+          WindowsFileAttributes.SYSTEM
+        ],
       );
 
       expect(updatedFileInfo.filePath, 'C:/test/file.txt');
@@ -67,8 +73,8 @@ void main() {
       expect(updatedFileInfo.fileSize, '1 KB');
       expect(updatedFileInfo.dwFileAttributes,
           FILE_FLAGS_AND_ATTRIBUTES.FILE_ATTRIBUTE_HIDDEN);
-      expect(updatedFileInfo.attributes,
-          [FileAttributes.HIDDEN, FileAttributes.SYSTEM]);
+      expect(updatedFileInfo.winAttributes,
+          [WindowsFileAttributes.HIDDEN, WindowsFileAttributes.SYSTEM]);
     });
 
     test('FileInfo toString returns correct string representation', () {
@@ -83,16 +89,20 @@ void main() {
         sizeBytes: 1024,
         fileSize: '1 KB',
         dwFileAttributes: FILE_FLAGS_AND_ATTRIBUTES.FILE_ATTRIBUTE_READONLY,
-        attributes: const [FileAttributes.READ_ONLY, FileAttributes.HIDDEN],
+        winAttributes: const [
+          WindowsFileAttributes.READ_ONLY,
+          WindowsFileAttributes.HIDDEN
+        ],
       );
 
       expect(
         fileInfo.toString(),
-        'FileInfo(filePath: C:/test/file.txt, fileName: file.txt, fileExtension: .txt, '
+        'FileMetadata(filePath: C:/test/file.txt, fileName: file.txt, fileExtension: .txt, '
         'fileType: Text Document, creationTime: 2023-08-10 00:00:00.000, '
         'modifiedTime: 2023-08-12 00:00:00.000, accessedTime: 2023-08-13 00:00:00.000, '
         'sizeBytes: 1024, fileSize: 1 KB, dwFileAttributes: ${FILE_FLAGS_AND_ATTRIBUTES.FILE_ATTRIBUTE_READONLY}, '
-        'attributes: [FileAttributes.READ_ONLY, FileAttributes.HIDDEN])',
+        'winAttributes: [WindowsFileAttributes.READ_ONLY, WindowsFileAttributes.HIDDEN], '
+        'androidAttributes: null)',
       );
     });
 
@@ -108,7 +118,7 @@ void main() {
         sizeBytes: 1024,
         fileSize: '1 KB',
         dwFileAttributes: FILE_FLAGS_AND_ATTRIBUTES.FILE_ATTRIBUTE_READONLY,
-        attributes: const [FileAttributes.READ_ONLY],
+        winAttributes: const [WindowsFileAttributes.READ_ONLY],
       );
 
       final fileInfo2 = FileMetadata(
@@ -122,7 +132,7 @@ void main() {
         sizeBytes: 1024,
         fileSize: '1 KB',
         dwFileAttributes: FILE_FLAGS_AND_ATTRIBUTES.FILE_ATTRIBUTE_READONLY,
-        attributes: const [FileAttributes.READ_ONLY],
+        winAttributes: const [WindowsFileAttributes.READ_ONLY],
       );
 
       final fileInfo3 = FileMetadata(
@@ -136,7 +146,7 @@ void main() {
         sizeBytes: 1024,
         fileSize: '1 KB',
         dwFileAttributes: FILE_FLAGS_AND_ATTRIBUTES.FILE_ATTRIBUTE_HIDDEN,
-        attributes: const [FileAttributes.HIDDEN],
+        winAttributes: const [WindowsFileAttributes.HIDDEN],
       );
 
       expect(fileInfo1, equals(fileInfo2));